[#2210][#728] Post-review refactoring (thanks @jotomo !)
