+++ wear/src/main/java/info/nightscout/androidaps/complications/IobIconComplication.java
@@ -7,10 +7,10 @@ import android.support.wearable.complications.ComplicationText;
 import android.util.Log;
 
 import info.nightscout.androidaps.R;
-import info.nightscout.androidaps.data.DisplayRawData;
+import info.nightscout.androidaps.data.RawDisplayData;
 import info.nightscout.androidaps.interaction.utils.SmallestDoubleString;
 
-import static info.nightscout.androidaps.interaction.utils.DisplayFormat.MAX_SHORT_FIELD;
+import static info.nightscout.androidaps.interaction.utils.DisplayFormat.MAX_FIELD_LEN_SHORT;
 
 /*
  * Created by dlvoy on 2019-11-12
@@ -19,12 +19,12 @@ public class IobIconComplication extends BaseComplicationProviderService {
 
     private static final String TAG = IobIconComplication.class.getSimpleName();
 
-    public ComplicationData buildComplicationData(int dataType, DisplayRawData raw, PendingIntent complicationPendingIntent) {
+    public ComplicationData buildComplicationData(int dataType, RawDisplayData raw, PendingIntent complicationPendingIntent) {
 
         ComplicationData complicationData = null;
 
         if (dataType == ComplicationData.TYPE_SHORT_TEXT) {
-            final String iob = new SmallestDoubleString(raw.sIOB1, SmallestDoubleString.Units.USE).minimise(MAX_SHORT_FIELD);
+            final String iob = new SmallestDoubleString(raw.sIOB1, SmallestDoubleString.Units.USE).minimise(MAX_FIELD_LEN_SHORT);
 
             final ComplicationData.Builder builder = new ComplicationData.Builder(ComplicationData.TYPE_SHORT_TEXT)
                     .setShortText(ComplicationText.plainText(iob))
