+++ wear/src/main/java/info/nightscout/androidaps/complications/LongStatusFlippedComplication.java
@@ -5,9 +5,8 @@ import android.support.wearable.complications.ComplicationData;
 import android.support.wearable.complications.ComplicationText;
 import android.util.Log;
 
-import info.nightscout.androidaps.data.DisplayRawData;
+import info.nightscout.androidaps.data.RawDisplayData;
 import info.nightscout.androidaps.interaction.utils.DisplayFormat;
-import info.nightscout.androidaps.interaction.utils.SmallestDoubleString;
 
 /*
  * Created by dlvoy on 2019-11-12
@@ -16,7 +15,7 @@ public class LongStatusFlippedComplication extends BaseComplicationProviderServi
 
     private static final String TAG = LongStatusFlippedComplication.class.getSimpleName();
 
-    public ComplicationData buildComplicationData(int dataType, DisplayRawData raw, PendingIntent complicationPendingIntent) {
+    public ComplicationData buildComplicationData(int dataType, RawDisplayData raw, PendingIntent complicationPendingIntent) {
 
         ComplicationData complicationData = null;
