+++ wear/src/main/java/info/nightscout/androidaps/complications/UploaderBattery.java
@@ -8,7 +8,7 @@ import android.util.Log;
 
 import androidx.annotation.DrawableRes;
 import info.nightscout.androidaps.R;
-import info.nightscout.androidaps.data.DisplayRawData;
+import info.nightscout.androidaps.data.RawDisplayData;
 
 /*
  * Created by dlvoy on 2019-11-12
@@ -17,7 +17,7 @@ public class UploaderBattery extends BaseComplicationProviderService {
 
     private static final String TAG = UploaderBattery.class.getSimpleName();
 
-    public ComplicationData buildComplicationData(int dataType, DisplayRawData raw, PendingIntent complicationPendingIntent) {
+    public ComplicationData buildComplicationData(int dataType, RawDisplayData raw, PendingIntent complicationPendingIntent) {
 
         ComplicationData complicationData = null;
