+++ wear/src/main/java/info/nightscout/androidaps/data/ListenerService.java
@@ -513,14 +513,14 @@ public class ListenerService extends WearableListenerService implements GoogleAp
                     Intent messageIntent = new Intent();
                     messageIntent.setAction(Intent.ACTION_SEND);
                     messageIntent.putExtra("status", dataMap.toBundle());
-                    Persistence.storeDataMap(DisplayRawData.STATUS_PERSISTENCE_KEY, dataMap);
+                    Persistence.storeDataMap(RawDisplayData.STATUS_PERSISTENCE_KEY, dataMap);
                     LocalBroadcastManager.getInstance(this).sendBroadcast(messageIntent);
                 } else if (path.equals(BASAL_DATA_PATH)){
                     dataMap = DataMapItem.fromDataItem(event.getDataItem()).getDataMap();
                     Intent messageIntent = new Intent();
                     messageIntent.setAction(Intent.ACTION_SEND);
                     messageIntent.putExtra("basals", dataMap.toBundle());
-                    Persistence.storeDataMap(DisplayRawData.BASALS_PERSISTENCE_KEY, dataMap);
+                    Persistence.storeDataMap(RawDisplayData.BASALS_PERSISTENCE_KEY, dataMap);
                     LocalBroadcastManager.getInstance(this).sendBroadcast(messageIntent);
                 } else if (path.equals(NEW_PREFERENCES_PATH)){
                     dataMap = DataMapItem.fromDataItem(event.getDataItem()).getDataMap();
@@ -544,7 +544,7 @@ public class ListenerService extends WearableListenerService implements GoogleAp
                     Intent messageIntent = new Intent();
                     messageIntent.setAction(Intent.ACTION_SEND);
                     messageIntent.putExtra("data", dataMap.toBundle());
-                    Persistence.storeDataMap(DisplayRawData.DATA_PERSISTENCE_KEY, dataMap);
+                    Persistence.storeDataMap(RawDisplayData.DATA_PERSISTENCE_KEY, dataMap);
                     LocalBroadcastManager.getInstance(this).sendBroadcast(messageIntent);
                 }
             }
