+++ wear/src/test/java/info/nightscout/androidaps/testing/mockers/AndroidMocker.java
@@ -12,29 +12,25 @@ import static org.powermock.api.mockito.PowerMockito.mockStatic;
 
 public class AndroidMocker {
 
-    public static void mockBase64() {
+    public static void mockBase64() throws Exception {
         mockStatic(android.util.Base64.class);
-        try {
-            PowerMockito.when(android.util.Base64.class, "decode", anyString(), anyInt()).thenAnswer(invocation -> {
 
-                final String payload = invocation.getArgument(0);
-                try {
-                    return Base64.getDecoder().decode(payload);
-                } catch (java.lang.IllegalArgumentException ex) {
-                    return null;
-                }
-            });
+        PowerMockito.when(android.util.Base64.class, "decode", anyString(), anyInt()).thenAnswer(invocation -> {
 
-            PowerMockito.when(android.util.Base64.class, "encodeToString", any(), anyInt()).thenAnswer(invocation -> {
+            final String payload = invocation.getArgument(0);
+            try {
+                return Base64.getDecoder().decode(payload);
+            } catch (java.lang.IllegalArgumentException ex) {
+                return null;
+            }
+        });
 
-                final byte[] payload = invocation.getArgument(0);
-                return  Base64.getEncoder().encodeToString(payload);
+        PowerMockito.when(android.util.Base64.class, "encodeToString", any(), anyInt()).thenAnswer(invocation -> {
 
-            });
+            final byte[] payload = invocation.getArgument(0);
+            return  Base64.getEncoder().encodeToString(payload);
 
-        } catch (Exception e) {
-            Assert.fail("Unable to mock objects: " + e.getMessage());
-        }
+        });
     }
 
 }
