+++ wear/src/test/java/info/nightscout/androidaps/testing/mockers/RawDataMocker.java
@@ -1,14 +1,14 @@
 package info.nightscout.androidaps.testing.mockers;
 
-import info.nightscout.androidaps.data.DisplayRawData;
+import info.nightscout.androidaps.data.RawDisplayData;
 import info.nightscout.androidaps.interaction.utils.SafeParse;
 
 import static info.nightscout.androidaps.testing.mockers.WearUtilMocker.backInTime;
 
 public class RawDataMocker {
 
-    public static DisplayRawData rawSgv(String sgv, int m, String deltaString) {
-        DisplayRawData raw = new DisplayRawData();
+    public static RawDisplayData rawSgv(String sgv, int m, String deltaString) {
+        RawDisplayData raw = new RawDisplayData();
         raw.datetime = backInTime(0, 0, m, 0);
         raw.sDelta = deltaString;
         raw.sSgv = sgv;
@@ -34,30 +34,30 @@ public class RawDataMocker {
         return raw;
     }
 
-    public static DisplayRawData rawDelta(int m, String delta) {
-        DisplayRawData raw = new DisplayRawData();
+    public static RawDisplayData rawDelta(int m, String delta) {
+        RawDisplayData raw = new RawDisplayData();
         raw.datetime = backInTime(0, 0, m, 0);
         raw.sDelta = delta;
         return raw;
     }
 
-    public static DisplayRawData rawCobIobBr(String cob, String iob, String br) {
-        DisplayRawData raw = new DisplayRawData();
+    public static RawDisplayData rawCobIobBr(String cob, String iob, String br) {
+        RawDisplayData raw = new RawDisplayData();
         raw.sCOB2 = cob;
         raw.sIOB1 = iob;
         raw.sBasalRate = br;
         return raw;
     }
 
-    public static DisplayRawData rawIob(String iob, String iob2) {
-        DisplayRawData raw = new DisplayRawData();
+    public static RawDisplayData rawIob(String iob, String iob2) {
+        RawDisplayData raw = new RawDisplayData();
         raw.sIOB1 = iob;
         raw.sIOB2 = iob2;
         return raw;
     }
 
-    public static DisplayRawData rawCob(String cob) {
-        DisplayRawData raw = new DisplayRawData();
+    public static RawDisplayData rawCob(String cob) {
+        RawDisplayData raw = new RawDisplayData();
         raw.sCOB2 = cob;
         return raw;
     }
